made signInEmail test suites
